chore: try in base folder
