+++ components/Header.tsx
@@ -5,16 +5,16 @@ import { ENV } from "utils/env"
 
 import Button from "@/components/base/Button"
 import { Flex } from "@/components/base/Flex"
-import { useAuth } from "../../hooks/useAuth"
-import { useIsMobile } from "../../hooks/useIsMobile"
-import ConnectButton from "../ConnectButton"
-import { Copy, Exit } from "../assets/icons"
-import DropdownMenu from "../base/DropdownMenu"
-import { Stack } from "../base/Stack"
-import { Headline } from "../base/Typography"
-import { Zorb } from "../base/Zorb"
+import { useAuth } from "../hooks/useAuth"
+import { useIsMobile } from "../hooks/useIsMobile"
+import ConnectButton from "./ConnectButton"
 import styles from "./Header.module.css"
 import { Navigation } from "./Navigation"
+import { Copy, Exit } from "./assets/icons"
+import DropdownMenu from "./base/DropdownMenu"
+import { Stack } from "./base/Stack"
+import { Headline } from "./base/Typography"
+import { Zorb } from "./base/Zorb"
 
 const fadeIn: Variants = {
   initial: { opacity: 0 },
